add todos for BasicManager
