+++ tests/simulation/sim/sim_config.go
@@ -20,7 +20,7 @@ var (
 	FlagBlockSizeValue          int
 	FlagLeanValue               bool
 	FlagCommitValue             bool
-	FlagOnOperationValue        bool // TODO: Remove in favor of binary search for invariant violation
+	FlagOnOperationValue        bool
 	FlagAllInvariantsValue      bool
 
 	FlagEnabledValue     bool
